+++ CHANGELOG.md
@@ -151,6 +151,7 @@
 - Quality Control : Set default printing setting and update translation
 - INVOICE : Configure BIRT to generate an Excel and a Word file
 - KEYWORD : remove Keyword model.
+- Advanced Import : trim data before import.
 
 ## Bug Fixes
 - TeamTask : Resolve NPE on save of new team task.